refactor(ui): decouple productivity screen via ViewModel and empty state sub-widget
